METRON-1717 Renamed Storm Profiler package
