+++ metron-analytics/metron-profiler/src/main/java/org/apache/metron/profiler/storm/FixedFrequencyFlushSignal.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.apache.metron.profiler.bolt;
+package org.apache.metron.profiler.storm;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
