+++ metron-analytics/metron-profiler/src/test/java/org/apache/metron/profiler/storm/KafkaEmitterTest.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.apache.metron.profiler.bolt;
+package org.apache.metron.profiler.storm;
 
 import com.google.common.collect.ImmutableMap;
 import org.adrianwalker.multilinestring.Multiline;
