+++ metron-analytics/metron-profiler/src/test/java/org/apache/metron/profiler/storm/ProfileHBaseMapperTest.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.apache.metron.profiler.bolt;
+package org.apache.metron.profiler.storm;
 
 import org.apache.metron.common.configuration.profiler.ProfileConfig;
 import org.apache.metron.common.configuration.profiler.ProfileResult;
